Added functionality to Graph class, began Eulerian paths
Graph class is now a package that other classes can use, since i don’t
want to reinvent the wheel or ugly copy-paste code into another
project. It also reports the degree of any node in the graph (i.e. the
number of links it has with other nodes: (1, 1):(1, 2), (1, 1):(2, 1),
(1, 1):(3, 1) is a degree of 3 for (1, 1). Working on correctly
detecting if there’s loops from e.g. (1, 1):(1, 1)